Cleaning up import of commander module again
Previously had a pull to do this but messed it up, so reverted and making a new pull.
Changed the way the Commander module is imported to what is on the README for it.

Modified the sites option by using Commander's Variadic option instead of the passed custom function to split the sites into an array.