finish setting up batch oclif command
Needed to update tsconfig for oclif to pick up the files in the command
directory. I will also rename the index_2 file later, just needed to
make sure there was not a naming conflict.

Lastly, since the file that is executing the command has changed, the
reports were getting put in the `bin` directory, so I outputed to get
the root directory by resolving ./